Bullet Speed Set
"As a player, I want my bullets to move at a set speed"

Bullet speed has been set to a public speed. Speed value is just a dummy for now and needs to be confirmed by the group. Further confirmation is also needed on whether speed is affected by perspective